+++ src/main/java/lotto/domain/LottoMachine.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class LottoMachine {
     private static final int PRICE_OF_LOTTO = 1_000;
@@ -10,15 +10,15 @@ public class LottoMachine {
     private LottoMachine() {
     }
 
-    static LottoBundle getLottos(int numberOfLotto) {
-        List<Lotto> lottos = IntStream.range(0, numberOfLotto)
+    static LottoBundle getLottos(long numberOfLotto) {
+        List<Lotto> lottos = LongStream.range(0, numberOfLotto)
                 .mapToObj(i -> LottoGenerator.generate())
                 .collect(Collectors.toList());
 
         return new LottoBundle(lottos);
     }
 
-    public static LottoBundle buyLottos(int money) {
+    public static LottoBundle buyLottos(long money) {
         if (money < 0) {
             throw new IllegalArgumentException();
         }
@@ -26,7 +26,7 @@ public class LottoMachine {
         return getLottos(getNumberOfAffordableLottos(money));
     }
 
-    public static int getNumberOfAffordableLottos(int money) {
+    public static long getNumberOfAffordableLottos(long money) {
         return money / PRICE_OF_LOTTO;
     }
 }
